refactor: move ATR calculation to indicators
