+++ crypto_bot/indicators/atr.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+
+import pandas as pd
+
+from crypto_bot.utils.indicator_cache import cache_series
+
+
+def calc_atr(df: pd.DataFrame, window: int = 14) -> float:
+    """Calculate the Average True Range using cached values."""
+    lookback = window
+    recent = df.iloc[-(lookback + 1) :]
+    high_low = recent["high"] - recent["low"]
+    high_close = (recent["high"] - recent["close"].shift()).abs()
+    low_close = (recent["low"] - recent["close"].shift()).abs()
+    tr = pd.concat([high_low, high_close, low_close], axis=1).max(axis=1)
+    atr_series = tr.rolling(window).mean()
+    cached = cache_series(f"atr_{window}", df, atr_series, lookback)
+    return float(cached.iloc[-1])
